refactor: @DisplayName 어노테이션 추가 및 패키지 구조 변경
